refactor: share daemon job facts and separate status tests
