+++ crates/ctx-daemon-cli/src/daemon_status.rs
@@ -7,36 +7,32 @@ pub(super) use render::{
     render_daemon_prepare_uninstall_receipt, render_daemon_status_human, DaemonStatusView,
 };
 
+fn job_status(job: Option<&Value>) -> &str {
+    job.and_then(|job| job.get("status"))
+        .and_then(Value::as_str)
+        .unwrap_or("unknown")
+}
+
+fn job_error(job: Option<&Value>) -> Option<&str> {
+    job.and_then(|job| job.get("last_error"))
+        .and_then(Value::as_str)
+        .filter(|error| !error.is_empty())
+}
+
 pub(super) fn daemon_jobs_failure_message(
     core_refresh: Option<&Value>,
     semantic: Option<&Value>,
 ) -> Option<String> {
-    if let Some(error) = core_refresh
-        .and_then(|job| job.get("last_error"))
-        .and_then(Value::as_str)
-        .filter(|error| !error.is_empty())
-    {
+    if let Some(error) = job_error(core_refresh) {
         return Some(format!("Core refresh failed: {error}"));
     }
-    if core_refresh
-        .and_then(|job| job.get("status"))
-        .and_then(Value::as_str)
-        == Some("failed")
-    {
+    if job_status(core_refresh) == "failed" {
         return Some("Core refresh failed".to_owned());
     }
-    if let Some(error) = semantic
-        .and_then(|job| job.get("last_error"))
-        .and_then(Value::as_str)
-        .filter(|error| !error.is_empty())
-    {
+    if let Some(error) = job_error(semantic) {
         return Some(format!("semantic indexing failed: {error}"));
     }
-    if semantic
-        .and_then(|job| job.get("status"))
-        .and_then(Value::as_str)
-        == Some("failed")
-    {
+    if job_status(semantic) == "failed" {
         return Some("semantic indexing failed".to_owned());
     }
     None
